pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "cv-routes"
version = "0.1.0"
description = "Segmentation sémantique de routes sur ortho-photos IGN (FLAIR-INC ONNX)"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "Pillow>=10.0",
    "onnxruntime>=1.16",
    "rasterio>=1.3",
    "requests>=2.31",
    "geopandas>=0.14",
    "shapely>=2.0",
    "fiona>=1.9",
]

[project.optional-dependencies]
hf   = ["huggingface_hub>=0.20"]
dev  = ["pytest>=7", "pytest-cov"]
full = ["huggingface_hub>=0.20", "matplotlib>=3.8"]

[tool.setuptools.packages.find]
where   = ["."]
include = ["cv_routes*"]
